hardcoded_bot.py
import asyncio
import pandas as pd
import logging
import os
from datetime import datetime
from binance import AsyncClient
from telegram import Bot, Update
from telegram.ext import Application, CommandHandler, ContextTypes

# Жестко заданные значения (ЗАМЕНИТЕ НА СВОИ!)
TELEGRAM_TOKEN = "8545214216:AAGRz-jmD-2989hx8LuP43Svse9AEC-v1HI"  # Например: "1234567890:ABCdefGHIjkl..."
BINANCE_API_KEY = "lvyzhyX59Jksmwxb5MD6krkfIl3kfAVKWvXokWyGNAUER3KXPQiE0WvWVhH2SBjA"
BINANCE_SECRET = "zjMTHl6MIHLE1RZDsHR1aG29Pamgak2E01mntFAyXsqHPHPhGgCNuRNLVNMKPGo0" 
ADMIN_CHAT_ID = "1077455247"  # Например: "123456789"

# Настройка логирования
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class SimpleCryptoBot:
    def __init__(self):
        self.tg_bot = Bot(token=TELEGRAM_TOKEN)
        self.binance_client = None
        
    async def initialize(self):
        try:
            self.binance_client = await AsyncClient.create(
                BINANCE_API_KEY, 
                BINANCE_SECRET
            )
            logger.info("✅ Боты инициализированы!")
        except Exception as e:
            logger.error(f"❌ Ошибка инициализации: {e}")
    
    async def start_command(self, update: Update, context: ContextTypes.DEFAULT_TYPE):
        chat_id = update.effective_chat.id
        await context.bot.send_message(
            chat_id=chat_id,
            text="🚀 Crypto Scanner Bot запущен!\n\nНачинаю сканирование рынка...",
            parse_mode='Markdown'
        )
        logger.info(f"📨 Получена команда /start от {chat_id}")
        
    async def run(self):
        await self.initialize()
        
        application = Application.builder().token(TELEGRAM_TOKEN).build()
        application.add_handler(CommandHandler("start", self.start_command))
        
        logger.info("🤖 Запускаю Telegram бота...")
        await application.run_polling()

async def main():
    bot = SimpleCryptoBot()
    await bot.run()

if __name__ == "__main__":
    asyncio.run(main())