requirements.txt
python-binance==1.0.16
python-telegram-bot==20.7
pandas==2.0.3
numpy==1.24.3
ta==0.10.2
aiohttp==3.8.5
python-dotenv==1.0.0
schedule==1.2.0
websockets==12.0
